// src/types/filters.ts
export interface DealFilters {
  search?: string;
  status?: DealStatus | DealStatus[];
  type?: DealType;
  propertyType?: PropertyType | PropertyType[];
  clientId?: string;
  createdById?: string;
  propertyState?: string;
  propertyCity?: string;
  closingDateFrom?: Date;
  closingDateTo?: Date;
  createdAtFrom?: Date;
  createdAtTo?: Date;
}